// src/types/post-block-update-types.ts
export enum PostBlockActionType {
  PdfGeneratePresignedUploadUrl = "PDF_GENERATE_PRESIGNED_UPLOAD_URL",
  GeneratePresignedUploadUrl = "GENERATE_PRESIGNED_UPLOAD_URL",
  StopQueryGeneration = "STOP_QUERY_GENERATION",
  GenerateDescription = "GENERATE_DESCRIPTION",
  StopCodeGeneration = "STOP_CODE_GENERATION",
  EditResultVariable = "EDIT_RESULT_VARIABLE",
  AnswerPdfQuestion = "ANSWER_PDF_QUESTION",
  PaginateDataframe = "PAGINATE_DATAFRAME",
  RunPythonBlock = "RUN_PYTHON_BLOCK",
  RunChartBlock = "RUN_CHART_BLOCK",
  RunTableBlock = "RUN_TABLE_BLOCK",
  UnverifyQuery = "UNVERIFY_QUERY",
  FixPython = "AUTO_FIX_PYTHON",
  RunCsvBlock = "RUN_CSV_BLOCK",
  RunSqlBlock = "RUN_SQL_BLOCK",
  UpdateBlock = "UPDATE_BLOCK",
  WritePython = "WRITE_PYTHON",
  VerifyQuery = "VERIFY_QUERY",
  DownloadCsv = "DOWNLOAD_CSV",
  DownloadSql = "DOWNLOAD_SQL",
  FixSql = "AUTO_FIX_SQL",
  IndexPDF = "INDEX_PDF",
  WriteSql = "WRITE_SQL",
}
// ...
export type PatchProjectResponseAction =
  // BotConversation:
  | CreateBotConversationMessageResponseAction
  | DeleteBotConversationMessageResponseAction
  | UpdateBotConversationMessageResponseAction
  | UpdateBotConversationAction
  | CreatePlanStep
  | UpdatePlanStep
  | CreatePlan
  | UpdatePlan
  // Project:
  | UpdateProjectResponseAction
  // Block:
  | CreateBlockResponseAction
  | DeleteBlockResponseAction
  | UpdateBlockResponseAction;

export type BasePostBlockActionOutput = {
  notebook_updates?: PatchProjectResponseAction[];
  error?: string | null;
  sources?: unknown[];
  answer?: string;
};
export type PostBlockResponse<T> = {
  action_output: BasePostBlockActionOutput & T;
  timestamp: ISODateString;
};